requirements.txt
python-dotenv
flask
flask-sqlalchemy
flask-restful
flask-marshmallow
psycopg2
flask-jwt-extended